Push already validated changes (#12)
* Feature/2024.0.0 update (#11)

* Updating to 2024.0.0 release

* Fixing run_benchmarks.py -h command malfunction

* Removing validation label from output removal command

* Removing empty cell

* Updating to pyuthon 3.10

* Update shell  type

* Removing kaggle installation instructions

* Updating python version to 3.10
